+++ src/main/java/com/sesiones/sesiones_backend/repository/CompetenciaRepository.java
@@ -7,9 +7,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sesiones.sesiones_backend.entity.Competencia;
 
-public interface CompetenciaRepository extends JpaRepository<Competencia, Long> {
+public interface CompetenciaRepository extends JpaRepository<Competencia, Integer> {
 
-    Optional<Competencia> findByIdAndAreaId(Long id, Long areaId);
+    Optional<Competencia> findByIdAndAreaId(Integer id, Integer areaId);
 
-    List<Competencia> findByIdIn(List<Long> ids);
+    List<Competencia> findByIdIn(List<Integer> ids);
 }
+
