+++ src/main/java/com/sesiones/sesiones_backend/repository/InstitucionRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sesiones.sesiones_backend.entity.Institucion;
 
-public interface InstitucionRepository extends JpaRepository<Institucion, Long> {
+public interface InstitucionRepository extends JpaRepository<Institucion, Integer> {
 
     Optional<Institucion> findByNombreIgnoreCase(String nombre);
 }
+
